+++ changelog.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `feature` Lector Role has option to select fav entity.
 - `security` CanConfig depends also on entityArray.
 - `social` Standar item-social code for updating.
+- `route` Fix problem navigating go same component.
 
 ### Update